Beverages 타입 추가, VendingMachine 내부 addStock, IncreaseAmount, showCurrentAmount 생성
