+++ VendingMachineApp/VendingMachineApp/Model/Beverages.swift
@@ -1,0 +1,27 @@
+//
+//  Beverages.swift
+//  VendingMachineApp
+//
+//  Created by sonjuhyeong on 2021/03/03.
+//
+
+import Foundation
+
+class Beverages {
+    
+    private var beverages: [Beverage]
+    
+    init() {
+        self.beverages = [Beverage]()
+    }
+    
+    func addBeverage(beverage: Beverage) {
+        beverages.append(beverage)
+    }
+    
+    func retrieveBeverage(completion: (Beverage) -> Void) {
+        beverages.forEach{(
+            completion($0)
+        )}
+    }
+}
